PR corrections message handler + sync.map + latency as summary
