+++ gbus/abstractions.go
@@ -4,9 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"github.com/sirupsen/logrus"
-	"reflect"
-	"runtime"
-	"strings"
 	"time"
 
 	"github.com/streadway/amqp"
@@ -109,15 +106,6 @@ type HandlerRegister interface {
 	HandleEvent(exchange, topic string, event Message, handler MessageHandler) error
 }
 
-//MessageHandler signature for all command handlers
-type MessageHandler func(invocation Invocation, message *BusMessage) error
-
-func (mg MessageHandler) Name() string {
-	funName := runtime.FuncForPC(reflect.ValueOf(mg).Pointer()).Name()
-	splits := strings.Split(funName, ".")
-	fn := strings.Replace(splits[len(splits)-1], "-fm", "", -1)
-	return fn
-}
 //Saga is the base interface for all Sagas.
 type Saga interface {
 	//StartedBy returns the messages that when received should create a new saga instance
